Fix short_interest_scan.py float parsing
FINRA Reg SHO daily files now report ShortVolume and TotalVolume as
floating-point numbers (market-maker hedging fractional share
accounting), not integers. The previous int() conversion silently
raised ValueError on every row, causing the scan to report "0 days"
for every ticker despite successfully fetching 25 days of data.
Switch to float() parsing.

+++ scripts/short_interest_scan.py
@@ -40,8 +40,9 @@ def parse_day(text, tickers_set):
             symbol = parts[1].strip().upper()
             if symbol not in tickers_set:
                 continue
-            short_vol = int(parts[2])
-            total_vol = int(parts[4])
+            # FINRA 2025+ 文件里 ShortVolume/TotalVolume 是浮点数（lot-split 做市商对冲）
+            short_vol = float(parts[2])
+            total_vol = float(parts[4])
             if total_vol > 0:
                 result[symbol] = (short_vol, total_vol)
         except (ValueError, IndexError):
